Working on Standardizing Gallery Elements.
Lots of work done to make it so that the gallery elements are more consistent.  So far zoomable and gallerific are ready.

Also added a new gallery element which gets the gallery and calls the right gallery type element.  Therefore a call to ->Element('gallery'), actually does all of the work, as long as you provide the model, and foreign key of the gallery.  So the whole call would look like this...

echo ->Element('gallery', array('model' => Gallery, 'foreignKey' => 33), array('plugin' => 'galleries'));

This standardization makes it so that all gallery element calls have to be updated though and they should no longer make use of the gallery id anywhere.  Always, the model and foreign key.