Made it so that Categories View actually imports the associated model, and uses variables from that model.
ClassRegistry::init() doens't work like it did before the 2.0 upgrade.  You used to be able to just give the model name, and even if the model was in a plugin it would find it.  That doesn't work now, but Cake if it can't find a model pretends that it can if the db table exists.   (ie. CatalogItem model can be inferred from the catalog_items table).

This leads to variables that you set in the model (like order) not being used if you don't actually load the model file though.

So specify the plugin, and a helpful way to do that is to use ZuhaInflector::pluginize(MODEL NAME), as that should return the plugin name for any model in the system nice and easily.

For example... App::uses(, ZuhaInflector::pluginize().'.Model');
 = new ;