Finally figured out what was causing the test to fail.
IMPORTANT NOTE : The test was asking for the default database connection, because in the GalleryAppModel we did not use the proper __construct function.  The __construct function passes the ds, datasource and because it wasn't passed, any test that was related to galleries would have had this problem.   Tough to find.

Also fixed the metable behavior so that find(first) is usable again, but only if you use the callbacks option set to false.